PMM-9849 moved run all check button added paddings
