+++ public/app/percona/check/components/FailedChecksTab/FailedChecksTab.messages.ts
@@ -1,6 +1,4 @@
 export const Messages = {
-  runDbChecks: 'Run Checks',
   showAll: 'Show all',
-  checksExecutionStarted: 'Running checks in the background... The results will be displayed here soon.',
   noChecks: 'No failed checks found',
 };
