+++ public/app/percona/check/components/FailedChecksTab/FailedChecksTab.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useEffect, useState, useCallback, useMemo } from 'react';
-import { LoaderButton, logger } from '@percona/platform-core';
+import { logger } from '@percona/platform-core';
 import { Cell, Column, Row } from 'react-table';
 import { useCancelToken } from 'app/percona/shared/components/hooks/cancelToken.hook';
 import { isApiCancelError } from 'app/percona/shared/helpers/api';
@@ -11,14 +11,12 @@ import { Spinner, useStyles2 } from '@grafana/ui';
 import { Messages } from './FailedChecksTab.messages';
 import { getStyles } from './FailedChecksTab.styles';
 import { stripServiceId } from './FailedChecksTab.utils';
-import { appEvents } from '../../../../core/app_events';
-import { AppEvents } from '@grafana/data';
+
 import { GET_ACTIVE_ALERTS_CANCEL_TOKEN } from './FailedChecksTab.constants';
 import { locationService } from '@grafana/runtime';
 
 export const FailedChecksTab: FC = () => {
   const [fetchAlertsPending, setFetchAlertsPending] = useState(true);
-  const [runChecksPending, setRunChecksPending] = useState(false);
   const [data, setData] = useState<FailedCheckSummary[]>([]);
   const styles = useStyles2(getStyles);
   const [generateToken] = useCancelToken();
@@ -61,18 +59,6 @@ export const FailedChecksTab: FC = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const handleRunChecksClick = async () => {
-    setRunChecksPending(true);
-    try {
-      await CheckService.runDbChecks();
-      appEvents.emit(AppEvents.alertSuccess, [Messages.checksExecutionStarted]);
-    } catch (e) {
-      logger.error(e);
-    } finally {
-      setRunChecksPending(false);
-    }
-  };
-
   const getRowProps = (row: Row<FailedCheckSummary>): ExtendedTableRowProps => ({
     key: row.original.serviceId,
     className: styles.row,
@@ -91,20 +77,7 @@ export const FailedChecksTab: FC = () => {
   }, []);
 
   return (
-    <>
-      <div className={styles.header}>
-        <div className={styles.actionButtons} data-testid="db-check-panel-actions">
-          <LoaderButton
-            type="button"
-            size="md"
-            loading={runChecksPending}
-            onClick={handleRunChecksClick}
-            className={styles.runChecksButton}
-          >
-            {Messages.runDbChecks}
-          </LoaderButton>
-        </div>
-      </div>
+    <div className={styles.contentWrapper}>
       <AlertsReloadContext.Provider value={{ fetchAlerts }}>
         {fetchAlertsPending ? (
           <div className={styles.spinner} data-testid="db-checks-failed-checks-spinner">
@@ -122,6 +95,6 @@ export const FailedChecksTab: FC = () => {
           />
         )}
       </AlertsReloadContext.Provider>
-    </>
+    </div>
   );
 };
